ported from 2.8

+++ arangod/Replication/InitialSyncer.cpp
@@ -570,8 +570,7 @@ int InitialSyncer::applyCollectionDump(
 ////////////////////////////////////////////////////////////////////////////////
 
 int InitialSyncer::handleCollectionDump(
-    arangodb::Transaction* trx, std::string const& cid,
-    TRI_transaction_collection_t* trxCollection,
+    TRI_vocbase_col_t* col, std::string const& cid,
     std::string const& collectionName, TRI_voc_tick_t maxTick,
     std::string& errorMsg) {
   std::string appendix;
@@ -611,8 +610,7 @@ int InitialSyncer::handleCollectionDump(
     url += "&chunkSize=" + StringUtils::itoa(chunkSize);
 
     std::string const typeString =
-        (trxCollection->_collection->_collection->_info.type() ==
-                 TRI_COL_TYPE_EDGE
+        (col->_type == TRI_COL_TYPE_EDGE
              ? "edge"
              : "document");
 
@@ -759,8 +757,31 @@ int InitialSyncer::handleCollectionDump(
     }
 
     if (res == TRI_ERROR_NO_ERROR) {
-      res = applyCollectionDump(trx, trxCollection, response.get(),
-                                markersProcessed, errorMsg);
+      SingleCollectionWriteTransaction<UINT64_MAX> trx(new StandaloneTransactionContext(), _vocbase, col->_cid);
+      
+      res = trx.begin();
+
+      if (res != TRI_ERROR_NO_ERROR) {
+        errorMsg = "unable to start transaction: " + std::string(TRI_errno_string(res));
+
+        return res;
+      }
+      
+      TRI_transaction_collection_t* trxCollection = trx.trxCollection();
+
+      if (trxCollection == nullptr) {
+        res = TRI_ERROR_INTERNAL;
+        errorMsg = "unable to start transaction: " + std::string(TRI_errno_string(res));
+      }
+            
+      if (trx.orderDitch(trx.trxCollection()) == nullptr) {
+        res = TRI_ERROR_OUT_OF_MEMORY;
+      }
+
+      if (res == TRI_ERROR_NO_ERROR) {
+        res = applyCollectionDump(&trx, trxCollection, response.get(), markersProcessed, errorMsg);
+        res = trx.finish(res);
+      }
     }
 
     if (res != TRI_ERROR_NO_ERROR) {
@@ -792,7 +813,7 @@ int InitialSyncer::handleCollectionDump(
 ////////////////////////////////////////////////////////////////////////////////
 
 int InitialSyncer::handleCollectionSync(
-    std::string const& cid, SingleCollectionWriteTransaction<UINT64_MAX>& trx,
+    TRI_vocbase_col_t* col, std::string const& cid, 
     std::string const& collectionName, TRI_voc_tick_t maxTick,
     std::string& errorMsg) {
   sendExtendBatch();
@@ -948,23 +969,39 @@ int InitialSyncer::handleCollectionSync(
   }
 
   if (countJson->_value._number <= 0.0) {
-    int res = trx.truncate(false);
+    // remote collection has no documents. now truncate our local collection
+    SingleCollectionWriteTransaction<UINT64_MAX> trx(new StandaloneTransactionContext(), _vocbase, col->_cid);
+
+    int res = trx.begin();
+
+    if (res != TRI_ERROR_NO_ERROR) {
+      errorMsg = std::string("unable to start transaction: ") + TRI_errno_string(res);
+
+      return res;
+    }
+      
+    if (trx.orderDitch(trx.trxCollection()) == nullptr) {
+      return TRI_ERROR_OUT_OF_MEMORY;
+    }
+
+    res = trx.truncate(false);
 
     if (res != TRI_ERROR_NO_ERROR) {
       errorMsg = "unable to truncate collection '" + collectionName + "': " +
                  TRI_errno_string(res);
-
       return res;
     }
 
-    return TRI_ERROR_NO_ERROR;
+    res = trx.commit();
+    
+    return res;
   }
 
   // now we can fetch the complete chunk information from the master
   int res;
 
   try {
-    res = handleSyncKeys(id, cid, trx, collectionName, maxTick, errorMsg);
+    res = handleSyncKeys(col, id, cid, collectionName, maxTick, errorMsg);
   } catch (arangodb::basics::Exception const& ex) {
     res = ex.code();
   } catch (...) {
@@ -978,16 +1015,14 @@ int InitialSyncer::handleCollectionSync(
 /// @brief incrementally fetch data from a collection
 ////////////////////////////////////////////////////////////////////////////////
 
-int InitialSyncer::handleSyncKeys(
+int InitialSyncer::handleSyncKeys(TRI_vocbase_col_t* col,
     std::string const& keysId, std::string const& cid,
-    SingleCollectionWriteTransaction<UINT64_MAX>& trx,
     std::string const& collectionName, TRI_voc_tick_t maxTick,
     std::string& errorMsg) {
   TRI_doc_update_policy_t policy(TRI_DOC_UPDATE_LAST_WRITE, 0, nullptr);
-  auto shaper = trx.documentCollection()->getShaper();
 
   bool const isEdge =
-      (trx.documentCollection()->_info.type() == TRI_COL_TYPE_EDGE);
+      (col->_type == TRI_COL_TYPE_EDGE);
 
   std::string progress =
       "collecting local keys for collection '" + collectionName + "'";
@@ -995,11 +1030,53 @@ int InitialSyncer::handleSyncKeys(
 
   // fetch all local keys from primary index
   std::vector<TRI_df_marker_t const*> markers;
+    
+  TRI_document_collection_t* document = nullptr;
+  ReplicationDitch* ditch = nullptr;
 
-  auto idx = trx.documentCollection()->primaryIndex();
-  markers.reserve(idx->size());
+  // acquire a replication ditch so no datafiles are thrown away from now on
+  // note: the ditch also protects against unloading the collection
+  {    
+    SingleCollectionReadOnlyTransaction trx(new StandaloneTransactionContext(), _vocbase, col->_cid);
+  
+    int res = trx.begin();
+  
+    if (res != TRI_ERROR_NO_ERROR) {
+      errorMsg = std::string("unable to start transaction: ") + TRI_errno_string(res);
+      return res;
+    }
+    
+    document = trx.documentCollection();
+    ditch = document->ditches()->createReplicationDitch(__FILE__, __LINE__);
+
+    if (ditch == nullptr) {
+      return TRI_ERROR_OUT_OF_MEMORY;
+    }
+  }
+
+  TRI_ASSERT(document != nullptr);
+  TRI_ASSERT(ditch != nullptr);
+
+  TRI_DEFER(document->ditches()->freeDitch(ditch));
 
   {
+    SingleCollectionReadOnlyTransaction trx(new StandaloneTransactionContext(), _vocbase, col->_cid);
+  
+    int res = trx.begin();
+  
+    if (res != TRI_ERROR_NO_ERROR) {
+      errorMsg = std::string("unable to start transaction: ") + TRI_errno_string(res);
+      return res;
+    }
+    
+    if (trx.orderDitch(trx.trxCollection()) == nullptr) {
+      return TRI_ERROR_OUT_OF_MEMORY;
+    }
+
+    auto idx = trx.documentCollection()->primaryIndex();
+    markers.reserve(idx->size());
+
+    uint64_t iterations = 0;
     arangodb::basics::BucketPosition position;
 
     uint64_t total = 0;
@@ -1014,7 +1091,20 @@ int InitialSyncer::handleSyncKeys(
       void const* marker = ptr->getDataPtr();
       auto df = static_cast<TRI_df_marker_t const*>(marker);
       markers.emplace_back(df);
+
+      if (++iterations % 100000 == 0) {
+        if (checkAborted()) {
+          return TRI_ERROR_REPLICATION_APPLIER_STOPPED;
+        }
+      }
     }
+
+    if (checkAborted()) {
+      return TRI_ERROR_REPLICATION_APPLIER_STOPPED;
+    }
+  
+    sendExtendBatch();
+    sendExtendBarrier();
 
     std::string progress = "sorting " + std::to_string(markers.size()) +
                            " local key(s) for collection '" + collectionName +
@@ -1090,6 +1180,19 @@ int InitialSyncer::handleSyncKeys(
   // remove all keys that are below first remote key or beyond last remote key
   if (n > 0) {
     // first chunk
+    SingleCollectionWriteTransaction<UINT64_MAX> trx(new StandaloneTransactionContext(), _vocbase, col->_cid);
+  
+    int res = trx.begin();
+  
+    if (res != TRI_ERROR_NO_ERROR) {
+      errorMsg = std::string("unable to start transaction: ") + TRI_errno_string(res);
+      return res;
+    }
+    
+    if (trx.orderDitch(trx.trxCollection()) == nullptr) {
+      return TRI_ERROR_OUT_OF_MEMORY;
+    }
+
     auto chunk = static_cast<TRI_json_t const*>(
         TRI_AtVector(&(json.get()->_value._objects), 0));
 
@@ -1130,6 +1233,8 @@ int InitialSyncer::handleSyncKeys(
                                              (TRI_voc_key_t)key, 0, nullptr,
                                              &policy, false, false);
     }
+
+    trx.commit();
   }
 
   size_t nextStart = 0;
@@ -1139,6 +1244,22 @@ int InitialSyncer::handleSyncKeys(
     if (checkAborted()) {
       return TRI_ERROR_REPLICATION_APPLIER_STOPPED;
     }
+
+    SingleCollectionWriteTransaction<UINT64_MAX> trx(new StandaloneTransactionContext(), _vocbase, col->_cid);
+  
+    int res = trx.begin();
+  
+    if (res != TRI_ERROR_NO_ERROR) {
+      errorMsg = std::string("unable to start transaction: ") + TRI_errno_string(res);
+      return res;
+    }
+    
+    if (trx.orderDitch(trx.trxCollection()) == nullptr) {
+      return TRI_ERROR_OUT_OF_MEMORY;
+    }
+    
+    auto idx = trx.documentCollection()->primaryIndex();
+    auto shaper = trx.documentCollection()->getShaper();
 
     size_t const currentChunkId = i;
     progress = "processing keys chunk " + std::to_string(currentChunkId) +
@@ -1534,6 +1655,12 @@ int InitialSyncer::handleSyncKeys(
         }
       }
     }
+
+    res = trx.commit();
+    
+    if (res != TRI_ERROR_NO_ERROR) {
+      return res;
+    }
   }
 
   return TRI_ERROR_NO_ERROR;
@@ -1557,6 +1684,22 @@ int InitialSyncer::changeCollection(TRI_vocbase_col_t* col,
   } catch (...) {
     return TRI_ERROR_INTERNAL;
   }
+}
+ 
+////////////////////////////////////////////////////////////////////////////////
+/// @brief determine the number of documents in a collection
+////////////////////////////////////////////////////////////////////////////////
+
+int64_t InitialSyncer::getSize(TRI_vocbase_col_t* col) {
+  SingleCollectionReadOnlyTransaction trx(new StandaloneTransactionContext(), _vocbase, col->_cid);
+
+  int res = trx.begin();
+
+  if (res != TRI_ERROR_NO_ERROR) {
+    return -1;
+  }
+   
+  return static_cast<int64_t>(trx.documentCollection()->size());
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -1653,6 +1796,10 @@ int InitialSyncer::handleCollection(TRI_json_t const* parameters,
 
             return res;
           }
+    
+          if (trx.orderDitch(trx.trxCollection()) == nullptr) {
+            return TRI_ERROR_OUT_OF_MEMORY;
+          }
 
           res = trx.truncate(false);
 
@@ -1746,38 +1893,11 @@ int InitialSyncer::handleCollection(TRI_json_t const* parameters,
     {
       READ_LOCKER(readLocker, _vocbase->_inventoryLock);
 
-      SingleCollectionWriteTransaction<UINT64_MAX> trx(
-          new StandaloneTransactionContext(), _vocbase, col->_cid);
-
-      res = trx.begin();
-
-      if (res != TRI_ERROR_NO_ERROR) {
-        errorMsg = "unable to start transaction: " +
-                   std::string(TRI_errno_string(res));
-
-        return res;
+      if (incremental && getSize(col) > 0) {
+        res = handleCollectionSync(col, StringUtils::itoa(cid), masterName, _masterInfo._lastLogTick, errorMsg);
       }
-
-
-      TRI_transaction_collection_t* trxCollection = trx.trxCollection();
-
-      if (trxCollection == nullptr) {
-        res = TRI_ERROR_INTERNAL;
-        errorMsg = "unable to start transaction: " +
-                   std::string(TRI_errno_string(res));
-      } else {
-        if (trx.orderDitch(trxCollection) == nullptr) {
-          return TRI_ERROR_OUT_OF_MEMORY;
-        }
-
-        if (incremental && trx.documentCollection()->size() > 0) {
-          res = handleCollectionSync(StringUtils::itoa(cid), trx, masterName,
-                                     _masterInfo._lastLogTick, errorMsg);
-        } else {
-          res = handleCollectionDump(&trx, StringUtils::itoa(cid),
-                                     trxCollection, masterName,
-                                     _masterInfo._lastLogTick, errorMsg);
-        }
+      else {
+        res = handleCollectionDump(col, StringUtils::itoa(cid), masterName, _masterInfo._lastLogTick, errorMsg);
       }
 
       if (res == TRI_ERROR_NO_ERROR) {
@@ -1786,11 +1906,27 @@ int InitialSyncer::handleCollection(TRI_json_t const* parameters,
         VPackValueLength const n = indexes.length();
 
         if (n > 0) {
+          sendExtendBatch();
+          sendExtendBarrier();
+
           std::string const progress = "creating " + std::to_string(n) +
                                        " index(es) for " + collectionMsg;
           setProgress(progress);
 
           try {
+            SingleCollectionWriteTransaction<UINT64_MAX> trx(new StandaloneTransactionContext(), _vocbase, col->_cid);
+      
+            res = trx.begin();
+
+            if (res != TRI_ERROR_NO_ERROR) {
+              errorMsg = "unable to start transaction: " + std::string(TRI_errno_string(res));
+              return res;
+            }
+
+            if (trx.orderDitch(trx.trxCollection()) == nullptr) {
+              return TRI_ERROR_OUT_OF_MEMORY;
+            }
+      
             TRI_document_collection_t* document = trx.documentCollection();
             TRI_ASSERT(document != nullptr);
 
@@ -1826,6 +1962,8 @@ int InitialSyncer::handleCollection(TRI_json_t const* parameters,
                 }
               }
             }
+      
+            res = trx.finish(res);
           } catch (arangodb::basics::Exception const& ex) {
             res = ex.code();
           } catch (...) {
@@ -1833,8 +1971,6 @@ int InitialSyncer::handleCollection(TRI_json_t const* parameters,
           }
         }
       }
-
-      res = trx.finish(res);
     }
 
     return res;
